# internal/swagger/swagger.yaml
basePath: /v1
definitions:
...
  config.SelectedModel:
    properties:
      frequency_penalty:
        type: number
      max_tokens:
        description: Overrides the default model configuration.
        type: integer
      model:
        description: |-
          The model id as used by the provider API.
          Required.
        type: string
      presence_penalty:
        type: number
      provider:
        description: |-
          The model provider, same as the key/id used in the providers config.
          Required.
        type: string
      provider_options:
        additionalProperties: {}
        description: Override provider specific options.
        type: object
      reasoning_effort:
        description: Only used by models that use the openai provider and need this
          set.
        type: string
      temperature:
        type: number
      think:
        description: Used by anthropic models that can reason to indicate if the model
          should think.
        type: boolean
      top_k:
        type: integer
      top_p:
        type: number
    type: object
  config.SelectedModelType:
    enum:
    - large
    - small
    type: string
    x-enum-varnames:
    - SelectedModelTypeLarge
    - SelectedModelTypeSmall
...
  github_com_neelworx-cpu_F4RGE-CLI_internal_config.Scope:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - ScopeGlobal
    - ScopeWorkspace
...
  proto.APIKeyKind:
    enum:
    - string
    - oauth
    type: string
    x-enum-varnames:
    - APIKeyKindString
    - APIKeyKindOAuth
...
  proto.ConfigCompactRequest:
    properties:
      enabled:
        type: boolean
      scope:
        $ref: '#/definitions/github_com_neelworx-cpu_F4RGE-CLI_internal_config.Scope'
    type: object
  proto.ConfigModelRequest:
    properties:
      model:
        $ref: '#/definitions/config.SelectedModel'
      model_type:
        $ref: '#/definitions/config.SelectedModelType'
      scope:
        $ref: '#/definitions/github_com_neelworx-cpu_F4RGE-CLI_internal_config.Scope'
    type: object
  proto.ConfigProviderKeyRequest:
    properties:
      api_key:
        items:
          type: integer
        type: array
      kind:
        $ref: '#/definitions/proto.APIKeyKind'
      provider_id:
        type: string
      scope:
        $ref: '#/definitions/github_com_neelworx-cpu_F4RGE-CLI_internal_config.Scope'
    type: object
  proto.ConfigRefreshOAuthRequest:
    properties:
      provider_id:
        type: string
      scope:
        $ref: '#/definitions/github_com_neelworx-cpu_F4RGE-CLI_internal_config.Scope'
    type: object
  proto.ConfigRemoveRequest:
    properties:
      key:
        type: string
      scope:
        $ref: '#/definitions/github_com_neelworx-cpu_F4RGE-CLI_internal_config.Scope'
    type: object
  proto.ConfigSetRequest:
    properties:
      key:
        type: string
      scope:
        $ref: '#/definitions/github_com_neelworx-cpu_F4RGE-CLI_internal_config.Scope'
      value: {}
    type: object
  proto.Error:
    properties:
      message:
        type: string
    type: object
...
  proto.ImportCopilotResponse:
    properties:
      success:
        type: boolean
      token: {}
    type: object
...
  proto.MCPClientInfo:
    properties:
      connected_at:
        type: string
      error: {}
      name:
        type: string
      prompt_count:
        type: integer
      resource_count:
        type: integer
      state:
        $ref: '#/definitions/proto.MCPState'
      tool_count:
        type: integer
    type: object
  proto.MCPGetPromptRequest:
    properties:
      args:
        additionalProperties:
          type: string
        type: object
      client_id:
        type: string
      prompt_id:
        type: string
    type: object
  proto.MCPGetPromptResponse:
    properties:
      prompt:
        type: string
    type: object
  proto.MCPNameRequest:
    properties:
      name:
        type: string
    type: object
  proto.MCPReadResourceRequest:
    properties:
      name:
        type: string
      uri:
        type: string
    type: object
  proto.MCPState:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - MCPStateDisabled
    - MCPStateStarting
    - MCPStateConnected
    - MCPStateError
...
  proto.ProjectInitPromptResponse:
    properties:
      prompt:
        type: string
    type: object
  proto.ProjectNeedsInitResponse:
    properties:
      needs_init:
        type: boolean
    type: object
...
info:
  contact:
    name: F4RGE
    url: https://4rged.app
  description: 4rged is a terminal-based AI coding assistant. This API is served over
    a Unix socket (or Windows named pipe) and provides programmatic access to workspaces,
    sessions, agents, LSP, MCP, and more.
  license:
    name: MIT
    url: https://github.com/neelworx-cpu/F4RGE-CLI/blob/main/LICENSE
  title: 4rged API
  version: "1.0"
paths:
...
  /workspaces/{id}/config/compact:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Config compact request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.ConfigCompactRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Set compact mode
      tags:
      - config
  /workspaces/{id}/config/import-copilot:
    post:
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.ImportCopilotResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Import Copilot credentials
      tags:
      - config
  /workspaces/{id}/config/model:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Config model request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.ConfigModelRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Set the preferred model
      tags:
      - config
  /workspaces/{id}/config/provider-key:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Config provider key request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.ConfigProviderKeyRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Set provider API key
      tags:
      - config
  /workspaces/{id}/config/refresh-oauth:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Refresh OAuth request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.ConfigRefreshOAuthRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Refresh OAuth token
      tags:
      - config
  /workspaces/{id}/config/remove:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Config remove request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.ConfigRemoveRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Remove a config field
      tags:
      - config
  /workspaces/{id}/config/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Config set request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.ConfigSetRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Set a config field
      tags:
      - config
...
  /workspaces/{id}/mcp/docker/disable:
    post:
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Disable Docker MCP
      tags:
      - mcp
  /workspaces/{id}/mcp/docker/enable:
    post:
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Enable Docker MCP
      tags:
      - mcp
  /workspaces/{id}/mcp/get-prompt:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP get prompt request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.MCPGetPromptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.MCPGetPromptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Get MCP prompt
      tags:
      - mcp
  /workspaces/{id}/mcp/read-resource:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP read resource request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.MCPReadResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Read MCP resource
      tags:
      - mcp
  /workspaces/{id}/mcp/refresh-prompts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP name request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.MCPNameRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Refresh MCP prompts
      tags:
      - mcp
  /workspaces/{id}/mcp/refresh-resources:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP name request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.MCPNameRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Refresh MCP resources
      tags:
      - mcp
  /workspaces/{id}/mcp/refresh-tools:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP name request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/proto.MCPNameRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/proto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Refresh MCP tools
      tags:
      - mcp
  /workspaces/{id}/mcp/states:
    get:
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/proto.MCPClientInfo'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Get MCP client states
      tags:
      - mcp
...
  /workspaces/{id}/project/init:
    post:
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Mark project as initialized
      tags:
      - project
  /workspaces/{id}/project/init-prompt:
    get:
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.ProjectInitPromptResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Get project initialization prompt
      tags:
      - project
  /workspaces/{id}/project/needs-init:
    get:
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.ProjectNeedsInitResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/proto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/proto.Error'
      summary: Check if project needs initialization
      tags:
      - project
...
swagger: "2.0"